Model checkpoint bug (#95)
* nested hierarchy bug

* use pre-populated optimizers while loading model

* review comments

* resume_training overwrites contents in model_dir. support for resuming training from specified checkpoint dir

* fixed failing model_dir test case

* added test cases

* moving tests to the same file. still some config mixup issue

* added __init__ to MariusConfig class, prevents re-use of old config values during testing

* added one more test scenario

* added resume_from_checkpoint desc in full_schema.rst

+++ src/cpp/include/nn/optim.h
@@ -12,7 +12,7 @@
 class Optimizer {
 
 public:
-    int num_steps_;
+    int64_t num_steps_;
 
     torch::OrderedDict<std::string, torch::OrderedDict<std::string, torch::Tensor>> state_dict_;
     torch::OrderedDict<std::string, torch::Tensor> param_dict_;
